Revised visibility of structs
Tag and TagIter are now in the header (like multiboot header) submodule
