+++ src/lib.rs
@@ -1,5 +1,6 @@
 #![no_std]
 
+use header::{Tag, TagIter};
 pub use boot_loader_name::BootLoaderNameTag;
 pub use elf_sections::{ElfSectionsTag, ElfSection, ElfSectionIter, ElfSectionType, ElfSectionFlags};
 pub use elf_sections::{ELF_SECTION_WRITABLE, ELF_SECTION_ALLOCATED, ELF_SECTION_EXECUTABLE};
@@ -9,6 +10,7 @@ pub use module::{ModuleTag, ModuleIter};
 #[macro_use]
 extern crate bitflags;
 
+mod header;
 mod boot_loader_name;
 mod elf_sections;
 mod memory_map;
@@ -45,7 +47,7 @@ impl BootInformation {
     }
 
     pub fn module_tags(&self) -> ModuleIter {
-        ModuleIter::new(self.tags())
+        ModuleIter{ iter: self.tags() }
     }
 
     pub fn boot_loader_name_tag(&self) -> Option<&'static BootLoaderNameTag> {
@@ -68,35 +70,5 @@ impl BootInformation {
 
     fn tags(&self) -> TagIter {
         TagIter{current: &self.first_tag as *const _}
-    }
-}
-
-#[repr(C)]
-pub struct Tag {
-    typ: u32,
-    size: u32,
-    // tag specific fields
-}
-
-pub struct TagIter {
-    current: *const Tag,
-}
-
-impl Iterator for TagIter {
-    type Item = &'static Tag;
-
-    fn next(&mut self) -> Option<&'static Tag> {
-        match unsafe{&*self.current} {
-            &Tag{typ:0, size:8} => None, // end tag
-            tag => {
-                // go to next tag
-                let mut tag_addr = self.current as usize;
-                tag_addr += tag.size as usize;
-                tag_addr = ((tag_addr-1) & !0x7) + 0x8; //align at 8 byte
-                self.current = tag_addr as *const _;
-
-                Some(tag)
-            },
-        }
     }
 }
